requirements.txt
absl-py==2.1.0
astunparse==1.6.3
boltons==23.1.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
deprecation==2.1.0
dill==0.3.8
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.42.0
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
ml-dtypes==0.2.0
mpmath==1.3.0
natsort==8.4.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
piexif==1.1.3
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pysaliency==0.2.21
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schema==0.7.5
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.15.0
termcolor==2.4.0
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
